+++ R/utils_list.R
@@ -1,9 +1,0 @@
-#' Utils for lists
-#'
-#' @description Utils functions.
-#'
-#' @noRd
-
-get_field_as_vector <- function(x, field) {
-  lapply(x, \(y) y[[field]] %||% NA) |> unlist()
-}